Implementacion de puestos de trabajo completa. Alta, Modificacion, Visualizacion y Reportes.

+++ database/migrations/2024_11_01_030618_create_puestos_trabajos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('titulo_puesto')->unique();
             $table->string('descripcion_generica');
-            $table->float('sueldo_base', 2);
+            $table->decimal('sueldo_base', 999, 2);
             $table->unsignedBigInteger('id_departamento_trabajo');
             $table->foreign('id_departamento_trabajo')->references('id')->on('departamento_trabajos');
             $table->timestamps();
